requirements.txt
proxmoxer==2.0.1
requests==2.31.0
PyYAML==6.0.1
